Removed `txpool.rs` from `fuel-core-interfaces`. Moved all stuff regarding channels into `services/txpool/service.rs`. Moved all functionality from `Sender`(old type for `fuel-core-interfaces`) into `Service` in `fuel-core-txpool`.
