+++ crates/fuel-core/src/service/adapters.rs
@@ -1,77 +1,45 @@
 use crate::{
     database::Database,
-    executor::Executor,
     service::Config,
 };
 #[cfg(feature = "p2p")]
 use fuel_core_p2p::orchestrator::Service as P2pService;
-use fuel_core_relayer::ports::RelayerDb;
 #[cfg(feature = "relayer")]
 use fuel_core_relayer::RelayerSynced;
-use fuel_core_storage::{
-    transactional::StorageTransaction,
-    Result as StorageResult,
-};
-use fuel_core_txpool::types::Word;
+use fuel_core_txpool::Service;
 use fuel_core_types::{
-    blockchain::{
-        primitives,
-        primitives::BlockHeight,
-    },
-    fuel_tx::{
-        Receipt,
-        Transaction,
-    },
-    services::{
-        executor::{
-            ExecutionBlock,
-            Result as ExecutorResult,
-            UncommittedResult,
-        },
-        p2p::{
-            GossipsubMessageAcceptance,
-            TransactionGossipData,
-        },
+    blockchain::SealedBlock,
+    fuel_tx::Transaction,
+    services::p2p::{
+        GossipsubMessageAcceptance,
+        TransactionGossipData,
     },
 };
 use std::sync::Arc;
-use tokio::{
-    sync::oneshot,
-    task::JoinHandle,
+use tokio::sync::{
+    broadcast::Receiver,
+    oneshot,
 };
+#[cfg(feature = "p2p")]
+use tokio::task::JoinHandle;
 
 pub mod poa;
+pub mod producer;
 pub mod txpool;
+
+/// This is used to get block import events from coordinator source
+/// and pass them to the txpool.
+pub struct BlockImportAdapter {
+    rx: Receiver<SealedBlock>,
+}
+
+pub struct TxPoolAdapter {
+    pub service: Arc<Service>,
+}
 
 pub struct ExecutorAdapter {
     pub database: Database,
     pub config: Config,
-}
-
-#[async_trait::async_trait]
-impl fuel_core_producer::ports::Executor<Database> for ExecutorAdapter {
-    fn execute_without_commit(
-        &self,
-        block: ExecutionBlock,
-    ) -> ExecutorResult<UncommittedResult<StorageTransaction<Database>>> {
-        let executor = Executor {
-            database: self.database.clone(),
-            config: self.config.clone(),
-        };
-        executor.execute_without_commit(block)
-    }
-
-    fn dry_run(
-        &self,
-        block: ExecutionBlock,
-        utxo_validation: Option<bool>,
-    ) -> ExecutorResult<Vec<Vec<Receipt>>> {
-        let executor = Executor {
-            database: self.database.clone(),
-            config: self.config.clone(),
-        };
-        executor.dry_run(block, utxo_validation)
-    }
 }
 
 pub struct MaybeRelayerAdapter {
@@ -80,52 +48,8 @@ pub struct MaybeRelayerAdapter {
     pub relayer_synced: Option<RelayerSynced>,
 }
 
-#[async_trait::async_trait]
-impl fuel_core_producer::ports::Relayer for MaybeRelayerAdapter {
-    async fn get_best_finalized_da_height(
-        &self,
-    ) -> StorageResult<primitives::DaBlockHeight> {
-        #[cfg(feature = "relayer")]
-        {
-            if let Some(sync) = self.relayer_synced.as_ref() {
-                sync.await_synced().await?;
-            }
-        }
-
-        Ok(self
-            .database
-            .get_finalized_da_height()
-            .await
-            .unwrap_or_default())
-    }
-}
-
 pub struct PoACoordinatorAdapter {
     pub block_producer: Arc<fuel_core_producer::Producer<Database>>,
-}
-
-#[async_trait::async_trait]
-impl fuel_core_poa::ports::BlockProducer<Database> for PoACoordinatorAdapter {
-    async fn produce_and_execute_block(
-        &self,
-        height: BlockHeight,
-        max_gas: Word,
-    ) -> anyhow::Result<UncommittedResult<StorageTransaction<Database>>> {
-        self.block_producer
-            .produce_and_execute_block(height, max_gas)
-            .await
-    }
-
-    async fn dry_run(
-        &self,
-        transaction: Transaction,
-        height: Option<BlockHeight>,
-        utxo_validation: Option<bool>,
-    ) -> anyhow::Result<Vec<Receipt>> {
-        self.block_producer
-            .dry_run(transaction, height, utxo_validation)
-            .await
-    }
 }
 
 pub struct P2pAdapter {
