+++ crates/fuel-core/src/service/adapters/poa.rs
@@ -1,41 +1,61 @@
-use fuel_core_interfaces::txpool::{
-    Sender,
-    TxPoolMpsc,
+use crate::{
+    database::Database,
+    service::adapters::{
+        PoACoordinatorAdapter,
+        TxPoolAdapter,
+    },
 };
 use fuel_core_poa::ports::TransactionPool;
+use fuel_core_storage::transactional::StorageTransaction;
 use fuel_core_types::{
-    fuel_tx::TxId,
-    services::txpool::ArcPoolTx,
+    blockchain::primitives::BlockHeight,
+    fuel_asm::Word,
+    fuel_tx::{
+        Receipt,
+        Transaction,
+        TxId,
+    },
+    services::{
+        executor::UncommittedResult,
+        txpool::ArcPoolTx,
+    },
 };
-use tokio::sync::oneshot;
-
-pub struct TxPoolAdapter {
-    pub(crate) sender: Sender,
-}
 
 #[async_trait::async_trait]
 impl TransactionPool for TxPoolAdapter {
     async fn pending_number(&self) -> anyhow::Result<usize> {
-        let (response, receiver) = oneshot::channel();
-        self.sender
-            .send(TxPoolMpsc::PendingNumber { response })
-            .await?;
-        receiver.await.map_err(Into::into)
+        self.service.pending_number().await
     }
 
     async fn total_consumable_gas(&self) -> anyhow::Result<u64> {
-        let (response, receiver) = oneshot::channel();
-        self.sender
-            .send(TxPoolMpsc::ConsumableGas { response })
-            .await?;
-        receiver.await.map_err(Into::into)
+        self.service.total_consumable_gas().await
     }
 
-    async fn remove_txs(&mut self, ids: Vec<TxId>) -> anyhow::Result<Vec<ArcPoolTx>> {
-        let (response, receiver) = oneshot::channel();
-        self.sender
-            .send(TxPoolMpsc::Remove { ids, response })
-            .await?;
-        receiver.await.map_err(Into::into)
+    async fn remove_txs(&self, ids: Vec<TxId>) -> anyhow::Result<Vec<ArcPoolTx>> {
+        self.service.remove_txs(ids).await
+    }
+}
+
+#[async_trait::async_trait]
+impl fuel_core_poa::ports::BlockProducer<Database> for PoACoordinatorAdapter {
+    async fn produce_and_execute_block(
+        &self,
+        height: BlockHeight,
+        max_gas: Word,
+    ) -> anyhow::Result<UncommittedResult<StorageTransaction<Database>>> {
+        self.block_producer
+            .produce_and_execute_block(height, max_gas)
+            .await
+    }
+
+    async fn dry_run(
+        &self,
+        transaction: Transaction,
+        height: Option<BlockHeight>,
+        utxo_validation: Option<bool>,
+    ) -> anyhow::Result<Vec<Receipt>> {
+        self.block_producer
+            .dry_run(transaction, height, utxo_validation)
+            .await
     }
 }
