+++ crates/services/consensus_module/poa/Cargo.toml
@@ -21,7 +21,6 @@ tokio = { version = "1.21", features = ["full"] }
 tracing = "0.1"
 
 [dev-dependencies]
-fuel-core-interfaces = { path = "../../../fuel-core-interfaces" }
 fuel-core-types = { path = "../../../types", features = ["test-helpers"] }
 mockall = "0.11"
 rand = "0.8"
