+++ crates/services/txpool/src/lib.rs
@@ -4,6 +4,7 @@ pub mod config;
 mod containers;
 pub mod ports;
 pub mod service;
+mod transaction_selector;
 pub mod txpool;
 pub mod types;
